Separate batch functionality from core
Currently there are a number of process scripts each of which acts as a
batch processor (e.g. taking a folder of data as input). Instead, I
will switch to a format where each function only acts on specific
files. There will be one general batch processing script which
implements the previous functionality in a more general way.