Bugfix & improvement to parameterised assistance function
Bugfix: grf & marker data is now being synchronised correctly at the end of the file

Improvement: the code now continues beyond the last full gait cycle, applying APO assistance for the last portion of the data even if it's incomplete. It's for the segment() function, used later on, to decide which data to keep or not.